requirements.txt
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
entsoe-py==0.7.5
exceptiongroup==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
numpy==2.2.6
pandas==2.3.3
pyTelegramBotAPI==4.12.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.4
pytz==2025.2
requests==2.32.5
schedule==1.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
